# swagger-output.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "AI-PMS.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/TeamPerformance":  {
                                               "post":  {
                                                            "tags":  [
                                                                         "TeamPerformance"
                                                                     ],
                                                            "requestBody":  {
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/TeamPerformanceRequest"
                                                                                                                                    }
                                                                                                                     },
                                                                                                "text/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/TeamPerformanceRequest"
                                                                                                                             }
                                                                                                              },
                                                                                                "application/*+json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/TeamPerformanceRequest"
                                                                                                                                      }
                                                                                                                       }
                                                                                            }
                                                                            },
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK"
                                                                                      }
                                                                          }
                                                        }
                                           },
                  "/api/TeamPerformance/project/{projectId}":  {
                                                                   "post":  {
                                                                                "tags":  [
                                                                                             "TeamPerformance"
                                                                                         ],
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "projectId",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string",
                                                                                                                      "format":  "uuid"
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "200":  {
                                                                                                              "description":  "OK"
                                                                                                          }
                                                                                              }
                                                                            }
                                                               }
              },
    "components":  {
                       "schemas":  {
...
                                       "TeamMemberPerformanceDto":  {
                                                                        "type":  "object",
                                                                        "properties":  {
                                                                                           "memberName":  {
                                                                                                              "type":  "string",
                                                                                                              "nullable":  true
                                                                                                          },
                                                                                           "totalTasks":  {
                                                                                                              "type":  "integer",
                                                                                                              "format":  "int32"
                                                                                                          },
                                                                                           "completedTasks":  {
                                                                                                                  "type":  "integer",
                                                                                                                  "format":  "int32"
                                                                                                              },
                                                                                           "inProgressTasks":  {
                                                                                                                   "type":  "integer",
                                                                                                                   "format":  "int32"
                                                                                                               },
                                                                                           "blockedTasks":  {
                                                                                                                "type":  "integer",
                                                                                                                "format":  "int32"
                                                                                                            },
                                                                                           "overdueTasks":  {
                                                                                                                "type":  "integer",
                                                                                                                "format":  "int32"
                                                                                                            },
                                                                                           "estimatedHours":  {
                                                                                                                  "type":  "integer",
                                                                                                                  "format":  "int32"
                                                                                                              },
                                                                                           "actualHours":  {
                                                                                                               "type":  "integer",
                                                                                                               "format":  "int32"
                                                                                                           },
                                                                                           "completionRate":  {
                                                                                                                  "type":  "number",
                                                                                                                  "format":  "double"
                                                                                                              }
                                                                                       },
                                                                        "additionalProperties":  false
                                                                    },
                                       "TeamPerformanceRequest":  {
                                                                      "type":  "object",
                                                                      "properties":  {
                                                                                         "projectName":  {
                                                                                                             "type":  "string",
                                                                                                             "nullable":  true
                                                                                                         },
                                                                                         "teamMembers":  {
                                                                                                             "type":  "array",
                                                                                                             "items":  {
                                                                                                                           "$ref":  "#/components/schemas/TeamMemberPerformanceDto"
                                                                                                                       },
                                                                                                             "nullable":  true
                                                                                                         }
                                                                                     },
                                                                      "additionalProperties":  false
                                                                  }
                                   }
                   },
...
}